More tests, plus tweaks for slopeTo()

+++ Week3/src/main/java/Point.java
@@ -79,10 +79,9 @@ public class Point implements Comparable<Point>
 			return Double.POSITIVE_INFINITY;
 		}
 		else
-		{
-			return ( that.y - this.y ) / ( that.x - this.x );
-		}
-
+        {
+            return (double) ( that.y - this.y ) / ( that.x - this.x );
+        }
 	}
 
     /**
